fix: Proxy Agent usage
The proxy variable will be used correctly. The HTTPS_PROXY / HTTP_PROXY
will be used, if PROTOLINT_PROXY is not set.

Evaluation takes place: if the remote url starts with https, HTTPS_PROXY
will be used, otherwise HTTP_PROXY will be used.
This interpretation is questionable, but makes sense in the first place.

Closes #426